Add ability to include a build tag
This enables conditional compiling. This is particularly useful for
shared protobufs for common APIs, where not all consumers have the same
preferences around build size vs benefits of vtprotobuf. This allows the
consumer of the library to chose whether vtprotobuf is used, rather than
only the producer of the library.

Additionally, it provides guidance to use `vtprotobuf` as the tag. The
hope is that if multiple libraries do this, a user can just set that tag
instead of `-tags=envoy_vt,enable_vtprotobuf_prometheus,vtproto_k8s` or
something like that.

See https://github.com/envoyproxy/envoy/pull/31172, where this will be
utilized (I think).