features/pool: clear repeated bytes/string field in ResetVT
Clearing repeated bytes/string field ensure that the memory can be
released by the garbage collector. Previously, the slice was only
resized to 0, but the underlying []byte/string were still referenced
by the slice.

String are immutable in Go and aren't supposed to be modified after they
are created, so reusing the allocated memory is not possible. The memory
allocated for []byte could be reused, but this requires some additional
modification to the generated Unmarshal code and would be incompatible
with the use of UnmarshalVTUnsafe. This could be considered in a
follow-up PR.

When repeating other types, we don't need to clear the slices:
- scalar: the slice itself hold the memory
- embedded messages: `.Reset` or `.ResetVT` is being called instead to
  clear the message.

`clear()` was introduced in Go 1.21, so the version has to be bumped.